Add files via upload
A COVID-19 cases dataset grouped by nation typically includes data on the number of confirmed cases, deaths, recoveries, and other relevant statistics related to the COVID-19 pandemic. Each dataset entry refers to a certain country or region.

The p-value obtained from the t-test (2.028114e-09) is extremely small. This suggests strong evidence against the null hypothesis that the population mean of the variable Deaths is equal to 10,000. Therefore, based on this test, we would reject the null hypothesis in favor of the alternative hypothesis that the population mean is different from 10,000.
The p-value obtained from the Hypothesis Testing by comparing a specific value is (8.695e-16) is also extremely small. This indicates strong evidence against the null hypothesis that the true probability of success/failure is 0.5. Hence, based on this test, we would reject the null hypothesis and conclude that the true probability is significantly different from 0.5.